users/app/requirements.txt
Flask
mysql-connector-python==8.0.17
requests
pandas
gunicorn
